autodoc: add a sphinx c:autodoc directive for the extracted doc
Add a sphinx extension for the c:autodoc directive which will
extract the doc and inject it into the document tree.

This part is based on Jani Nikula's project: hawkmoth [1]
which has exactly the same goal as this series but which
use clang to parse the C code and extract the bloc-coments.

[1] https://github.com/jnikula/hawkmoth

Based-on-work-by: Jani Nikula <jani@nikula.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>